# docs/api/openapi.v1.yaml
openapi: 3.0.3
info:
  title: MedTracker API
  version: v1
  description: |
    Hosted MedTracker API contract for mobile clients, API clients, and portable migration.
    API requests are limited per client IP to 300 requests in each five-minute window. GET
    household data exports are also limited to 10 per minute; POST sync batches to 30 per
    minute; GET medication lookup to 60 per minute; POST AI medication suggestions to 10 per
    minute; and GET household admin audit logs to 100 per minute. Limits use epoch-aligned
    fixed windows. The first exceeded limit rejects the request with 429 and Retry-After,
    RateLimit-Limit, RateLimit-Remaining and RateLimit-Reset response headers. Counters are
    local to each server process. Counter storage is bounded; a new client is rejected with
    429 while the store is full of live counters. Expired counters are reclaimed. Health
    checks and direct loopback clients are exempt in production. The client IP is the socket
    peer unless a deployment explicitly configures trusted proxy IPs through
    API_TRUSTED_PROXY_IPS. A configured loopback proxy is not exempt, and untrusted
    forwarding headers do not set the client IP.
    Ordinary household operations accept a valid API session, household app token,
    or mobile OAuth bearer with the medtracker scope. SMART integration grants
    authorize SMART/FHIR access, not ordinary household operations. A household
...
    last_used_at at most once every five minutes.
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /households/{household_id}/medications:
    get:
      operationId: listMedications
      tags:
        - Household
        - Medications
      summary: List visible medications.
      parameters:
        - $ref: '#/components/parameters/household_id'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/per_page'
        - $ref: '#/components/parameters/updated_since'
      responses:
        '200':
          description: Medications collection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationCollectionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      operationId: createMedication
      tags:
        - Household
        - Medications
      summary: Create a medication.
      parameters:
        - $ref: '#/components/parameters/household_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationCreateRequest'
      responses:
        '201':
          description: Medication created.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /households/{household_id}/dosage_options:
    get:
      operationId: listDosageOptions
      tags:
        - Household
        - Dosage options
      summary: List medication dosage options.
      description: Active household members receive options only for medications they can read. Owners and administrators can read every medication in their household. Other members can read a medication linked to a person with an active view grant, or an unlinked medication they created. Options for medications outside that scope are omitted from the collection and its total count.
      parameters:
        - $ref: '#/components/parameters/household_id'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/per_page'
        - $ref: '#/components/parameters/updated_since'
      responses:
        '200':
          description: Dosage option collection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DosageOptionCollectionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      operationId: createDosageOption
      tags:
        - Household
        - Dosage options
      summary: Create a medication dosage option.
      description: Household owners and administrators may create options for medications in their household. Amount must be greater than zero; minimum hours, current supply, and reorder threshold must be nonnegative; unit and frequency must contain non-whitespace text. Only one option per medication may be the adult default and only one may be the child default; a duplicate returns 422 without changing either option. Creation clears the parent medication's single-dose amount and refreshes its sync version. If the option has current supply, the parent inventory is recalculated from all options with current supply, including their reorder thresholds and last-restock baseline. The option and parent changes are atomic.
      parameters:
        - $ref: '#/components/parameters/household_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DosageOptionCreateRequest'
      responses:
        '201':
          description: Dosage option created.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DosageOptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /households/{household_id}/medication_takes:
    get:
      operationId: listMedicationTakes
      tags:
        - Household
        - Medication takes
      summary: List visible medication takes.
      parameters:
        - $ref: '#/components/parameters/household_id'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/per_page'
        - $ref: '#/components/parameters/updated_since'
      responses:
        '200':
          description: Medication takes collection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationTakeCollectionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      operationId: createMedicationTake
      tags:
        - Household
        - Medication takes
      summary: Record a medication take.
      parameters:
        - $ref: '#/components/parameters/household_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationTakeCreateRequest'
      responses:
        '200':
          description: Existing medication take returned for a repeated client UUID.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationTakeResponse'
        '201':
          description: Medication take created.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationTakeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Household API sessions, household app tokens, and mobile OAuth medtracker access tokens authorize ordinary household operations. SMART integration grants do not.
  parameters:
...
    household_id:
      name: household_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/NumericId'
...
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    per_page:
      name: per_page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    updated_since:
      name: updated_since
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Timestamp'
...
  headers:
...
    etag:
      description: Version identifier for optimistic concurrency.
      required: true
      schema:
        type: string
    retry_after:
      description: Seconds until the request may be retried.
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
...
    BadRequest:
      description: A required request parameter is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: The bearer credential is absent, invalid, expired, revoked, or no longer active.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: The credential cannot perform the operation in the routed household or person scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: The resource does not exist inside the authorized scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationFailed:
      description: The submitted attributes or filters are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimited:
      description: The request exceeded an API rate limit.
      headers:
        Retry-After:
          $ref: '#/components/headers/retry_after'
        RateLimit-Limit:
          description: Maximum requests allowed by the matched limit.
          schema:
            type: integer
            minimum: 1
        RateLimit-Remaining:
          description: Requests remaining in the matched window.
          schema:
            type: integer
            enum: [0]
        RateLimit-Reset:
          description: Unix timestamp in seconds when the matched window resets.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitErrorEnvelope'
  schemas:
    NumericId:
      type: integer
      minimum: 1
    PortableId:
      type: string
      format: uuid
...
    Timestamp:
      type: string
      format: date-time
...
    DecimalValue:
      type: string
      pattern: '^-?[0-9]+(?:\.[0-9]+)?$'
...
    PaginationMeta:
      type: object
      additionalProperties: false
      required:
        - page
        - per_page
        - total_count
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 100
        total_count:
          type: integer
          minimum: 0
...
    Medication:
      type: object
      additionalProperties: false
      required:
        - id
        - portable_id
        - name
        - display_name
        - category
        - description
        - dose_amount
        - dose_unit
        - current_supply
        - reorder_threshold
        - reorder_status
        - location_id
        - location_portable_id
        - updated_at
        - low_stock
        - out_of_stock
        - days_until_low_stock
        - days_until_out_of_stock
      properties:
        id:
          $ref: '#/components/schemas/NumericId'
        portable_id:
          $ref: '#/components/schemas/PortableId'
        name:
          type: string
          minLength: 1
        display_name:
          type: string
          minLength: 1
        friendly_name:
          type: string
          nullable: true
        barcode:
          type: string
          nullable: true
        warnings:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        dose_amount:
          $ref: '#/components/schemas/NullableDecimalValue'
        dose_unit:
          type: string
          nullable: true
        current_supply:
          $ref: '#/components/schemas/NullableDecimalValue'
        reorder_threshold:
          $ref: '#/components/schemas/DecimalValue'
        reorder_status:
          type: string
          nullable: true
          enum:
            - ordered
            - received
        location_id:
          $ref: '#/components/schemas/NumericId'
        location_portable_id:
          $ref: '#/components/schemas/NullablePortableId'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
        low_stock:
          type: boolean
        out_of_stock:
          type: boolean
        days_until_low_stock:
          type: integer
          nullable: true
        days_until_out_of_stock:
          type: integer
          nullable: true
    MedicationResponse:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Medication'
    MedicationCollectionResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Medication'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    MedicationCreateRequest:
      type: object
      additionalProperties: false
      required:
        - medication
      properties:
        medication:
          $ref: '#/components/schemas/MedicationCreateAttributes'
...
    MedicationCreateAttributes:
      type: object
      additionalProperties: false
      required:
        - name
        - reorder_threshold
        - location_id
      properties:
        name:
          type: string
          minLength: 1
        friendly_name:
          type: string
        barcode:
          type: string
          pattern: '^[0-9]{13,14}$'
        dmd_code:
          type: string
        dmd_system:
          type: string
        dmd_concept_class:
          type: string
        category:
          type: string
        description:
          type: string
        dose_amount:
          $ref: '#/components/schemas/NullableDecimalValue'
        dose_unit:
          type: string
          enum:
            - tablet
            - capsule
            - gummy
            - mg
            - ml
            - g
            - mcg
            - IU
            - spray
            - drop
            - sachet
            - pad
        current_supply:
          $ref: '#/components/schemas/NullableDecimalValue'
        reorder_threshold:
          $ref: '#/components/schemas/DecimalValue'
        warnings:
          type: string
        location_id:
          $ref: '#/components/schemas/NumericId'
        default_schedule_type:
          type: string
          enum:
            - daily
            - multiple_daily
            - weekly
            - specific_dates
            - prn
            - tapering
            - every_other_day
...
    DosageOption:
      type: object
      additionalProperties: false
      required:
        - id
        - portable_id
        - medication_id
        - medication_portable_id
        - amount
        - unit
        - frequency
        - description
        - default_for_adults
        - default_for_children
        - default_max_daily_doses
        - default_min_hours_between_doses
        - default_dose_cycle
        - current_supply
        - reorder_threshold
        - updated_at
      properties:
        id:
          $ref: '#/components/schemas/NumericId'
        portable_id:
          $ref: '#/components/schemas/PortableId'
        medication_id:
          $ref: '#/components/schemas/NumericId'
        medication_portable_id:
          $ref: '#/components/schemas/PortableId'
        amount:
          $ref: '#/components/schemas/DecimalValue'
        unit:
          type: string
          minLength: 1
        frequency:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        default_for_adults:
          type: boolean
        default_for_children:
          type: boolean
        default_max_daily_doses:
          type: integer
          minimum: 1
        default_min_hours_between_doses:
          $ref: '#/components/schemas/DecimalValue'
        default_dose_cycle:
          type: string
          enum:
            - daily
            - weekly
            - monthly
        current_supply:
          $ref: '#/components/schemas/NullableDecimalValue'
        reorder_threshold:
          $ref: '#/components/schemas/NullableDecimalValue'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    DosageOptionResponse:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/DosageOption'
    DosageOptionCollectionResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DosageOption'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    DosageOptionCreateRequest:
      type: object
      additionalProperties: false
      required:
        - dosage_option
      properties:
        dosage_option:
          $ref: '#/components/schemas/DosageOptionCreateAttributes'
...
    DosageOptionCreateAttributes:
      type: object
      additionalProperties: false
      required:
        - medication_id
        - amount
        - unit
        - frequency
        - default_max_daily_doses
        - default_min_hours_between_doses
        - default_dose_cycle
      properties:
        medication_id:
          $ref: '#/components/schemas/ResourceIdentifier'
        amount:
          $ref: '#/components/schemas/DecimalValue'
        unit:
          type: string
          minLength: 1
        frequency:
          type: string
          minLength: 1
        description:
          type: string
        default_for_adults:
          type: boolean
        default_for_children:
          type: boolean
        default_max_daily_doses:
          type: integer
          minimum: 1
        default_min_hours_between_doses:
          $ref: '#/components/schemas/DecimalValue'
        default_dose_cycle:
          type: string
          enum:
            - daily
            - weekly
            - monthly
        current_supply:
          $ref: '#/components/schemas/NullableDecimalValue'
        reorder_threshold:
          $ref: '#/components/schemas/NullableDecimalValue'
...
    MedicationTake:
      type: object
      additionalProperties: false
      required:
        - id
        - portable_id
        - client_uuid
        - schedule_id
        - schedule_portable_id
        - person_medication_id
        - person_medication_portable_id
        - taken_from_medication_id
        - taken_from_medication_portable_id
        - taken_from_location_id
        - taken_from_location_portable_id
        - dose_amount
        - dose_unit
        - taken_at
        - updated_at
        - person_id
        - person_portable_id
        - medication_id
        - medication_portable_id
      properties:
        id:
          $ref: '#/components/schemas/NumericId'
        portable_id:
          $ref: '#/components/schemas/PortableId'
        client_uuid:
          $ref: '#/components/schemas/NullablePortableId'
        schedule_id:
          $ref: '#/components/schemas/NullableNumericId'
        schedule_portable_id:
          $ref: '#/components/schemas/NullablePortableId'
        person_medication_id:
          $ref: '#/components/schemas/NullableNumericId'
        person_medication_portable_id:
          $ref: '#/components/schemas/NullablePortableId'
        taken_from_medication_id:
          $ref: '#/components/schemas/NullableNumericId'
        taken_from_medication_portable_id:
          $ref: '#/components/schemas/NullablePortableId'
        taken_from_location_id:
          $ref: '#/components/schemas/NullableNumericId'
        taken_from_location_portable_id:
          $ref: '#/components/schemas/NullablePortableId'
        dose_amount:
          $ref: '#/components/schemas/NullableDecimalValue'
        dose_unit:
          type: string
          nullable: true
        taken_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          $ref: '#/components/schemas/Timestamp'
        person_id:
          $ref: '#/components/schemas/NullableNumericId'
        person_portable_id:
          $ref: '#/components/schemas/NullablePortableId'
        medication_id:
          $ref: '#/components/schemas/NullableNumericId'
        medication_portable_id:
          $ref: '#/components/schemas/NullablePortableId'
    MedicationTakeResponse:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/MedicationTake'
    MedicationTakeCollectionResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MedicationTake'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    MedicationTakeCreateRequest:
      type: object
      additionalProperties: false
      required:
        - medication_take
      properties:
        medication_take:
          type: object
          additionalProperties: false
          required:
            - source_type
            - source_id
            - taken_at
          properties:
            client_uuid:
              $ref: '#/components/schemas/PortableId'
            source_type:
              type: string
              enum:
                - schedule
                - person_medication
            source_id:
              $ref: '#/components/schemas/ResourceIdentifier'
            taken_at:
              $ref: '#/components/schemas/Timestamp'
            dose_amount:
              $ref: '#/components/schemas/DecimalValue'
            dose_unit:
              type: string
              minLength: 1
            taken_from_medication_id:
              $ref: '#/components/schemas/NumericId'
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'
...
    RateLimitErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/RateLimitError'
...